refactor: backup_phrase module
