+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -2,7 +2,7 @@ import 'package:auto_route/auto_route.dart';
 
 import '../../../features/add_funds/presentation/add_funds_screen.dart';
 import '../../../features/app_lock/presentation/routes.dart';
-import '../../../features/backup_phrase/presentation/routes.dart';
+import '../../../features/backup_phrase/module.dart';
 import '../../../features/incoming_split_key_payments/presentation/routes.dart';
 import '../../../features/outgoing_direct_payments/presentation/routes.dart';
 import '../../../features/outgoing_split_key_payments/presentation/routes.dart';
