+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/backup_phrase_flow/backup_confirm_phrase_screen.dart
@@ -2,11 +2,11 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../l10n/l10n.dart';
-import '../../../../ui/app_bar.dart';
-import '../../../../ui/onboarding_screen.dart';
-import '../../../../ui/recovery_phrase_text_view.dart';
-import '../../../../ui/theme.dart';
+import '../../../../../l10n/l10n.dart';
+import '../../../../../ui/app_bar.dart';
+import '../../../../../ui/onboarding_screen.dart';
+import '../../../../../ui/recovery_phrase_text_view.dart';
+import '../../../../../ui/theme.dart';
 import 'backup_phrase_flow_screen.dart';
 import 'components/grid_phrase.dart';
