+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/backup_phrase_flow/backup_phrase_flow_screen.dart
@@ -2,7 +2,7 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../app/routes.dart';
+import '../../../../../app/routes.dart';
 import '../../bl/puzzle_reminder_bloc.dart';
 
 class BackupPhraseFlowScreen extends StatefulWidget {
