+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/backup_phrase_flow/backup_warning_screen.dart
@@ -2,11 +2,11 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../l10n/l10n.dart';
-import '../../../../ui/app_bar.dart';
-import '../../../../ui/info_widget.dart';
-import '../../../../ui/onboarding_screen.dart';
-import '../../../../ui/theme.dart';
+import '../../../../../l10n/l10n.dart';
+import '../../../../../ui/app_bar.dart';
+import '../../../../../ui/info_widget.dart';
+import '../../../../../ui/onboarding_screen.dart';
+import '../../../../../ui/theme.dart';
 import 'backup_phrase_flow_screen.dart';
 
 class BackupWarningScreen extends StatelessWidget {
