+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/backup_phrase_flow/components/grid_phrase.dart
@@ -2,7 +2,7 @@ import 'package:dfunc/dfunc.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../../ui/button.dart';
+import '../../../../../../ui/button.dart';
 
 class GridPhrase extends StatefulWidget {
   const GridPhrase({
