+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/puzzle_reminder/components/puzzle_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../ui/app_bar.dart';
-import '../../../../../ui/decorated_window/decorated_window_header.dart';
-import '../../../../../ui/theme.dart';
+import '../../../../../../ui/app_bar.dart';
+import '../../../../../../ui/decorated_window/decorated_window_header.dart';
+import '../../../../../../ui/theme.dart';
 
 class PuzzleScreen extends StatelessWidget {
   const PuzzleScreen({
