+++ packages/cryptoplease/lib/features/backup_phrase/src/presentation/puzzle_reminder/puzzle_reminder_message_screen.dart
@@ -1,10 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../app/routes.gr.dart';
-import '../../../../l10n/l10n.dart';
-import '../../../../ui/button.dart';
-import '../../../../ui/message_info_widget.dart';
+import '../../../../../app/routes.gr.dart';
+import '../../../../../l10n/l10n.dart';
+import '../../../../../ui/button.dart';
+import '../../../../../ui/message_info_widget.dart';
 import 'components/puzzle_screen.dart';
 
 class PuzzleReminderMessageScreen extends StatelessWidget {
