+++ packages/cryptoplease/lib/features/backup_phrase/src/routes.dart
@@ -1,12 +1,12 @@
 import 'package:auto_route/auto_route.dart';
 
-import 'backup_phrase_flow/backup_confirm_phrase_screen.dart';
-import 'backup_phrase_flow/backup_phrase_flow_screen.dart';
-import 'backup_phrase_flow/backup_phrase_screen.dart';
-import 'backup_phrase_flow/backup_phrase_success_screen.dart';
-import 'backup_phrase_flow/backup_warning_screen.dart';
-import 'puzzle_reminder/puzzle_reminder_message_screen.dart';
-import 'puzzle_reminder/puzzle_reminder_setup_screen.dart';
+import 'presentation/backup_phrase_flow/backup_confirm_phrase_screen.dart';
+import 'presentation/backup_phrase_flow/backup_phrase_flow_screen.dart';
+import 'presentation/backup_phrase_flow/backup_phrase_screen.dart';
+import 'presentation/backup_phrase_flow/backup_phrase_success_screen.dart';
+import 'presentation/backup_phrase_flow/backup_warning_screen.dart';
+import 'presentation/puzzle_reminder/puzzle_reminder_message_screen.dart';
+import 'presentation/puzzle_reminder/puzzle_reminder_setup_screen.dart';
 
 const List<AutoRoute<dynamic>> backupPhraseRoutes = [
   AutoRoute<void>(
